Fix Index type to i32, remove partial support for other types
Unfortunately, the AnyDSL runtime only supports i32 in certain areas,
such as alloc and parallel.

+++ include/anybem/types.h
@@ -5,7 +5,7 @@
 
 namespace anybem {
 	using real_t     = double;
-	using index_t    = int64_t;
+	using index_t    = int32_t;
 
 	struct Position {
 		real_t x;
